+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 
 ## [0.12.0] - Skipped
 
+### Added
+
+- Add getters for `Validator` object to expose more data to users.
+- Add test file for `Validator` object.
+
 ### [0.11.3] - 2024-12-10
 
 ### Added